feat: include path information if config parsing fails.
If there is an problem with parting config file, the error message does
not give information about context of the error, that is in which
option/path there is a problem.

Example of an error befor this commit:

> Error: relative URL without a base
>
> Location:
>     src/tree/config.rs:52:22

After commit

> Error: bridged_trees.optimism.provider.rpc_endpoint: relative URL without a base
>
> Location:
>     src/tree/config.rs:52:22

After the commit you could see which config option deserializer did not
like.